started combining query & mutation tests in one suite; added test for creating tags
