+++ lib/section.js
@@ -47,6 +47,10 @@ const HASURA_INSERT_TAG = `mutation insert_single_tag($locale_code: String, $tit
   insert_tags_one(object: {slug: $slug, published: $published, tag_translations: {data: {locale_code: $locale_code, title: $title}}}) {
     id
     slug
+    tag_translations {
+      locale_code
+      title
+    }
   }
 }`;
